Refactor repository & add tests
This commit refactoring the repository so each layer in the MVC would handle their own responsibilities.

Moreover, this commit also adding (finally) tests to the repository to ensure the quality of the code going to "production".

This is done as a first step before a refactoring that will allow us to use H2 as an in-memory db that we can use while the service running on heroku without paying for it.